pyproject.toml
[project]
name = 'League'
version = '0.1.0'
description = 'AWS Lambda & DynamoDB backed web application'
readme = 'README.md'
requires-python = '>=3.12'
dependencies = [
    'boto3>=1.42.0,<2.0.0',
    'botocore>=1.42.0,<2.0.0',
    'aws-cdk-lib>=2.244.0,<3.0.0',
    'constructs>=10.5.0,<11.0.0'
]

[tool.pytest.ini_options]
addopts = '-rA -s -v'
markers = [
    'sessions: Session related tests',
    'registration: Registration related tests',
    'login: Login related tests',
    'homepage: Homepage related tests'
]

[project.optional-dependencies]
dev = [
    'moto>=5.1,<6.0',
    'pytest>=9.0,<10.0',
    'pytest-uuid==1.0.0',
    'pytest-mock>=3.15,<4.0',
    'types-boto3[dynamodb]>=1.42,<2.0',
    'types-boto3[secretsmanager]',
    'botocore>=1.42.0,<2.0.0',
    'ruff>=0.15.10,<1.0.0',
    'mypy>=1.20.0,<2.0.0',
    'mypy_boto3_dynamodb>=1.42.0,<2.0.0'
]

[tool.ruff]
line-length = 79
target-version = 'py312'

[tool.ruff.lint]
extend-select = [
    'F',        # Pyflakes rules
    'W',        # PyCodeStyle warnings
    'E',        # PyCodeStyle errors
    'I',        # Sort imports properly
    'UP',       # Warn if certain things can changed due to newer Python versions
    'C4',       # Catch incorrect use of comprehensions, dict, list, etc
    'RET',      # Good return practices
    'SIM',      # Common simplification rules
    'TID',      # Some good import practices
    'TC',       # Enforce importing certain types in a TYPE_CHECKING block
    'PTH',      # Use pathlib instead of os.path
    'TD',       # Be diligent with TODO comments
]

[tool.ruff.format]
quote-style = 'single'
indent-style = 'space'

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
disallow_incomplete_defs = true